+++ src/packages/user/user-group/workspace/user-group/components/user-group-entity-user-permission-list.element.ts
