+++ src/packages/user/user-group/workspace/user-group/components/user-group-granular-permission-list.element.ts
