+++ src/packages/user/user-group/workspace/user-group/manifests.ts
@@ -1,0 +1,42 @@
+import type {
+	ManifestWorkspaces,
+	ManifestWorkspaceActions,
+	ManifestWorkspaceView,
+	ManifestTypes,
+} from '@umbraco-cms/backoffice/extension-registry';
+import { UmbSubmitWorkspaceAction } from '@umbraco-cms/backoffice/workspace';
+
+const workspace: ManifestWorkspaces = {
+	type: 'workspace',
+	kind: 'routable',
+	alias: 'Umb.Workspace.UserGroup',
+	name: 'User Group Workspace',
+	api: () => import('./user-group-workspace.context.js'),
+	meta: {
+		entityType: 'user-group',
+	},
+};
+
+const workspaceViews: Array<ManifestWorkspaceView> = [];
+const workspaceActions: Array<ManifestWorkspaceActions> = [
+	{
+		type: 'workspaceAction',
+		kind: 'default',
+		alias: 'Umb.WorkspaceAction.UserGroup.Save',
+		name: 'Save User Group Workspace Action',
+		api: UmbSubmitWorkspaceAction,
+		meta: {
+			label: '#buttons_save',
+			look: 'primary',
+			color: 'positive',
+		},
+		conditions: [
+			{
+				alias: 'Umb.Condition.WorkspaceAlias',
+				match: workspace.alias,
+			},
+		],
+	},
+];
+
+export const manifests: Array<ManifestTypes> = [workspace, ...workspaceViews, ...workspaceActions];
